Recentered the offset tracking
Offset is now tracked as if 0,0 was at the center of the image.
This change will hopefully make it more user friendly to display and let
user modify the offsets from the UI.

I also refactored some code, renaming variables surrounding the offsets to make it more
clear what it does.