Implemented flood fill mask modifier
The modifier samples a location on the image for color, and spreads mask
around to this and similar colors that don't pass the threshhold
tolerance.

This is implemented using the same method frame maker created frames,
the relevant function parts have been extracted and generalized so both,
frame maker and this modifier can use the same underlaying flood fill code.

Also fixed channel count bug in image resampling code that resurfaced
when used with grayscale images.