Reversed the order which modifiers are applied
This makes it so adding modifiers adds them earlier in the stack than
existing ones, having the ones on top in the UI be applied later. This
makes it consistent with how it usually works with layers in other
applications, and is more comfortable to add more modifiers since it
leads to fewer reordering.